requirements.txt
asgiref==3.4.1
astroid==2.9.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.9
django-allauth==0.47.0
django-binary-database-files==1.0.15
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-on-heroku==1.1.2
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
#djangorestframework-simplejwt==5.0.0
dynaconf==3.1.7
gunicorn==20.1.0
heroku==0.1.4
hvac==0.11.2
idna==3.3
importlib-metadata==4.10.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
#jwt==1.3.1
knox==0.1.14
lazy-object-proxy==1.7.1
loguru==0.5.3
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
Pillow==9.0.1
platformdirs==2.4.1
pycparser==2.21
psycopg2==2.8.6
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
python-dateutil==1.5
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
validators==0.18.2
whitenoise==6.0.0
win32-setctime==1.1.0
wrapt==1.13.3
zipp==3.7.0